refactor: 패키지 순환 구조 제거
