+++ backend/src/test/java/com/woowacourse/moamoa/auth/controller/interceptor/PathRequestMatcherTest.java
@@ -2,6 +2,7 @@ package com.woowacourse.moamoa.auth.controller.interceptor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowacourse.moamoa.auth.config.interceptor.PathRequestMatcher;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
