+++ backend/src/test/java/com/woowacourse/moamoa/common/WebMVCTest.java
@@ -6,8 +6,8 @@ import static org.mockito.Mockito.when;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.moamoa.alarm.SlackAlarmSender;
 import com.woowacourse.moamoa.alarm.SlackUsersClient;
-import com.woowacourse.moamoa.auth.controller.interceptor.PathMatcherContainer;
-import com.woowacourse.moamoa.auth.controller.interceptor.PathMatcherInterceptor;
+import com.woowacourse.moamoa.auth.config.interceptor.PathMatcherContainer;
+import com.woowacourse.moamoa.auth.config.interceptor.PathMatcherInterceptor;
 import com.woowacourse.moamoa.auth.infrastructure.GithubOAuthClient;
 import com.woowacourse.moamoa.auth.infrastructure.JwtTokenProvider;
 import com.woowacourse.moamoa.auth.infrastructure.TokenProvider;
